Extract some components
